\name{allTotallyUnimodular}
\alias{allTotallyUnimodular}
\title{Test if a list of matrices are all unimodular}
\usage{
  allTotallyUnimodular(L)
}
\arguments{
  \item{L}{A list of objects of class matrix.}
}
\value{
  logical vector of length \code{length(L)}
}
\description{
  Helper function for \code{\link{raghavachari}}
}
\seealso{
  \code{\link{isTotallyUnimodular}}
}
\keyword{internal}

