\name{raghavachari}
\alias{raghavachari}
\title{Determine if a matrix is unimodular using recursive Raghavachari criterium...}
\usage{raghavachari(A)
}
\description{Determine if a matrix is unimodular using recursive Raghavachari criterium}
\details{This function is \code{deducorrect} internal}
\value{\code{TRUE} or \code{FALSE}}
\seealso{\code{\link{isTotallyUnimodular}}}
\arguments{\item{A}{An object of class Matrix in \eqn{\{-1,0,1\}^{m\times n}}.}
}

