% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getUsername}
\alias{getUsername}
\title{Get name of R user.}
\usage{
getUsername()
}
\value{
The username
}
\description{
Get name of R user.
}
\details{
This function returns the username.

This function is not exported.
}
\keyword{internal}

