% Generated by roxygen2 (4.0.1): do not edit by hand
\name{newdeducorrect}
\alias{newdeducorrect}
\alias{print.deducorrect}
\title{Generate an S3 \code{deducorrect} object}
\usage{
newdeducorrect(corrected, corrections, status, Call = sys.call(-1))
}
\arguments{
\item{corrected}{The corrected data.frame}

\item{corrections}{A \code{data.frame} listing old and new values for every row and variable where corrections were applied}

\item{status}{A \code{data.frame} with at least one \code{\link{status}} column.}

\item{Call}{Optionally, a \code{call} object.}
}
\value{
an S3 object of class \code{deducorrect}
}
\description{
Generate an S3 \code{deducorrect} object
}
\seealso{
\code{\link{deducorrect-object}}
}

