\name{raghavachari}
\alias{raghavachari}
\title{Determine if a matrix is unimodular using recursive Raghavachari criterium}
\usage{
  raghavachari(A)
}
\arguments{
  \item{A}{An object of class Matrix in
  \eqn{\{-1,0,1\}^{m\times n}}.}
}
\value{
  \code{TRUE} or \code{FALSE}
}
\description{
  This function is \code{deducorrect} internal
}
\seealso{
  \code{\link{isTotallyUnimodular}}
}

